move table component
